requirements.txt
# Core dependencies for GUIDE pipeline

# Video download and processing
yt-dlp
av>=14.0.0
opencv-python>=4.7.0
imageio
imageio-ffmpeg
pillow>=12.0.0

# Speech recognition
openai-whisper>=20240930

# LLM / VLM APIs
openai>=1.60.0
langchain>=0.3.0
langchain-openai>=0.3.0

# Video RAG
faiss-cpu>=1.10.0
fake-useragent>=2.0.0

# Subtitle processing
webvtt-py>=0.5.0

# Image processing
scikit-image>=0.22.0
numpy
scipy>=1.15.0

# Deep learning (for local VLM models)
torch>=2.0.0
torchvision
torchaudio

# Utilities
python-dotenv>=1.0.0
requests>=2.30.0
tqdm>=4.60.0
pydantic>=2.0.0
PyYAML>=6.0

# Google API (for video search fallback)
google-api-python-client>=2.100.0

# Tokenization
tiktoken>=0.9.0

# Optional: Local Qwen VL models
# transformers>=4.40.0
# accelerate>=0.30.0
# qwen-vl-utils
